Type annotate the tokens vector
In order to get a reference to a token in the vector, the Rust
compiler needs to know the type, it can't be inferred. We'll need
a reference to a previously added token to change its jump address. When
we come accross ']', we'll have to go back to the corresponding '['
and change its jump address.